Fix rfswitch to handle timeouts
After a timeout, any subsequent request was getting the previous
response, which reported the wrong port, causing errors forever after a
timeout. Ignore valid messages with wrong ports, and only trigger an
error if a correct port is not found within the timeout allowed. There
is only ever one client to the rfswitch service, so this should be
sufficient but numbered requests and responses would provide additional
protection against false positives.