py(feat): add twoport-dataset-2
This dataset has good cal measurements and good result, and shows how the data
should be loaded into the TwelveTerm cal object. Repeat for
short/load/open: S11 result for the
reflective standard on port one becomes S11, S11 result for port two
becomes S22 in a two-port S-param set that has S12, S21 set to zero.
It's not clear that the S11, S22 three-term Cal substitution needs to be
done, based on the data from the good VNA that we have in
twoport-dataset-2. Needs more investigation.