py (feat): test protoc server
gRPC server implemented for two port calibration, and tested with python
client using data from CopperMountainVNA - using calibrated data, and
checking that the results are the same before and after calibration
(because the error correction will be basically zero). Does not protect
against an echo server implementation passing test, but for now we don't
have a ready-made dataset with the raw DUT values to use.